requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
ultralytics>=8.0.0
pillow>=10.0.0
pydantic>=2.0.0
